CODE: Refactor Makefile added compiler flags

+++ include/types/types.h
@@ -6,10 +6,10 @@ enum { size_buffer_default = 0xFFFF };
 typedef unsigned char      byte_t;
 typedef unsigned short int word_t;
 
-typedef int          bool_t;
-typedef FILE         file_t;
-typedef unsigned int address_word_t;
-typedef unsigned int address_byte_t;
+typedef int                bool_t;
+typedef FILE               file_t;
+typedef unsigned short int address_word_t;
+typedef unsigned short int address_byte_t;
 
 typedef unsigned int size_byte_buffer;
 typedef unsigned int size_word_buffer;
